Finished integrating the module installer with the admin interface.
Added a hook to allow modules to expand the installation form. An abstract class was also built to allow the easy handling of the return data.

A bug was fixed in the module installer that made it so it was not loading the extended installation class.

A new spl autoload function was added for abstracts. For php5.3 all these autoload functions will be converted down to one based on namespaces.