Fixed bug that was causing modules not to load by ID.
Finished building SEO friendly URLs. Still need to program URL class to handle them.

Created special form type for Locations, which will create a select field (for now) containing the appropriately filtered locations (so if you only want a certain type, such as directory or module, only those will show up in the select field). 

Added new function to Location class which returns an array with the ID and 'path' of each location.

Fixed file locking in the Cache filesystem handler. Now when the file is locked for writing the cache class will just move on instead of waiting for the lock to be released.