Config changes- 	Changed the names of some of the configuration variables to make them all consistent (some were plural, others singular, even though they all referred to a collection of classes). This meant changing all the areas where they were used as well as the edit action. 	Updated installer with changed name 	Added "javascript' variable to the path section of the config
Theme updates-
	Rewrote the way the theme file loads its data. Now when it first loads it checks three places for javascript files (each installed package, the selected theme's javascript folder, and finally the system's javascript loaded) to build an internal library of all javascript files in the system and a link to each of them. A similar system has been set up for CSS files, only there is no final system folder to look in (it checks the installed packages and then the theme).

For both the CSS and Javascript files a simple syntax has been put in place: {library.}name{.min}.extension

Links are retrieved from the theme class by passing the name of the file and an optional library name to a function (one for CSS files, another for Javascript ones). If a minified version of the file is available and that option is enabled the system will return that link, otherwise it will return the primary link.

The jquery core and ui plugin have been renamed to fit the current scheme (specifically, periods in the name section were replaced with underscores).

PackageList
	Added option to retrieve only packages with installed modules present.

Default
	Added forced login when 'guest' uses the admin engine.

ModuleInfo
	Added 'checkAuth" function to check permissions for the module