This commit is all theme related.
The Page class was updated to use information from the new Theme class to get its information. As part of this update the process for generating javascript and css includes has been changed- now the functions that tell the page class what to load only store the data, they don't load the links. Instead the links are generated as part of the 'runtimeProcessTemplate" function.

The Form class was altered to add a CSS file to the page class, which in turn uses the theme class to figure out which css file to include.

The Theme class was altered to load an addition folder for CSS (a system one). The pattern for "glob" was altered so only files, not directories, are loaded.