BentoBlog!
This commit includes the sql, classes and actions needed to install and run this first version of BentoBlog. 

Current version can add, edit, delete and list blog posts. There is currently no main display, and the features of the blog are very slim, but this is just the beginning of this one.