Placed 'cacheHandler' configuration variable under System.
Extended default cache time.

Added a cache added time to the saved information.

Added a function to extend the cache time of an item without recalculating it. (I sort of cheated on this one- it essentially just resaves the recalled data, which is the only way to do it with the filesystem handler but could probably be improved with other systems)

Changed error handling to the custom system.

Set the various cache function to use the START_TIME constant instead of running the microtime function each time its called.