Forms-
This update incorporates the server side validation rules. If the rules aren't met, the checkSubmit function returns false and if the form is outputted again it will trigger the jQuery form validation function to show the user which inputs need to be corrected.

New validation rules are also much easier to deal with.