RequestWrapper
This class is essentially the dispatcher. It loads the user request, routes it, and handles any errors.

This class has two types of helpers-

IOProcessors do exactly that. Each processor handles inputs in a unique way- through the command line, over HTTP (relying on the GET variables for action routing), or REST (which uses HTTP but also relies a bit more heavily on header information and the other fun features of the HTTP protocol for figuring out whats going on). These classes also handle actually outputting the results the RequestWrapper  sends them.

OutputControllers essentially handle all formating of data before it gets sent out. Currently we have output controllers for JSON, HTML and coming soon XML. A special controller is also set up for the web administration interface.

This separation gives us flexibility when adding new ways to interact with the system, as a new format will be available across all methods of interaction and any new method of interacting with the system will, theoretically, be accessible by applications that have not necessarily been built with any consideration for that method.