Weirdest bug ever-
$path was defined as the path to a file. When I tried to include it with caching disabled, it would load fine, but with caching enabled it wouldn't load even though the damn path variable was the same.  At the same time, include was still returning true, even though it was throwing warnings, but it wasn't including any of the code it included.

Putting "file_exists()" in front of the include solved the bug. This doesn't make any sense, since this in no way changes the $path variable in any way, and so if include was failing before not only should it still be failing (which its not) but so should file_exists, which is not. This also does nothing to explain why the bug only existed when caching was enabled.