Updated the http processor to properly respond to cache handlers sent by web clients. If the web client uses either/both of the 'if-modified-since' and 'if-none-match' headers to see if changes have been made.
The entire request is still run, and if even a single character is changed the etag values won't match. If it turns out the client has the freshest copy, a 304 header will be sent out and the content will be held back. 

This code only comes into effect for 'get' requests.