Created new ActionBase class for actions that do not deal with models.
This was a complete copy paste job, resulting in duplicate code and all that crap- I don't care, it was easier to copy it over and alter it until it worked with the different action types than it would have been to abstract things out before. Eventually this class will be the 'base' class, and the model action class will extend off of this, reducing that redundancy.