Update to the model support actions.
Altered the Base so that the checkAuth function no longer creates the permissions object, but rather calls it from another function. This allows other actions (Add, Edit) to edit the permissions object without having to repeat the rest of the code again.

Altered the Add action to extend off of the Base action.
Altered the Add action to check the parent location for add permissions, since the new model is extending off of that location.

Altered the Add action (specifically its admin view) to redirect users after successfully adding an item to the items admin read page.