Reintroduced the extended session id regeneration code. This code tries to solve a problem where, upon changing a session id and destroying the old session, a new page would be requested with the old ID (generally a page called by ajax methods very quickly after the page that triggered the id change, but before that page loaded and the browser could get the new id) which in turn would cause a third session to be created. This third, new (guest) session takes precedence because it is the last one received by the browser, and thus ends up logging the user out.
Here we set an expiration on the old session instead of just killing it, thus allowing those incoming connections to still work properly. The obsolete session will not generate a new id, but after its time is up will just get destroyed. The current 15 second grace period should be more than sufficient.

This probably won't be as big of an issue with hosted applications, since there is higher latency there and requests are more likely to be spread out more, but its still is feasible. Most applications seem to avoid this by just not destroying the old session, but that seems stupid- if you don't destroy the old session, you're just increasing the number of possible (successful) possibilities for someone trying to hijack a session.