This update includes a new feature, error logging. The constant ERROR_LOGGING, an integer, is used to define the severity of the error that would be logged (this is similar to the DEBUG constant, which controls error outputs to the display).
Since this is an expansion of the RequestLogging code, this code uses the same database methodology, allowing admins to use a separate database for logs.

As part of this update an ErrorLog call was placed in the RequestWrapper 'handleError' function for all but a select few exceptions, with the errorType 'BLOCKING' for easy identification in the database since these are most likely programming errors.