For real this time-
Fixed a bug in the PackageInfo class. If the packageinfo class loads a file but the file does not contain the appropriate class, an exception is thrown. Since the package info class is called by the model registry class, when the exception looks up the active site it ends up throwing an exception, which loops until php stops the script. This update fixes that.