Updated the password class, adding a method to check whether a stored password is using the current settings/protocol.
Updated the ActiveUser class to replace the stored password hash if it was stored using an old method. This check occurs whenever the user tries to log in, meaning that if we update the password hash function at any point we will have a mechanism in place to transparently switch from the old protocol to the new one.