This is a rather big update to the Model system.
Previously there was one type of model, that based on location. To allow us to manage all the date structures with the model system- for instance, editing Users- this update create new generic model classes and splits the location based models into their own subclass of those base models. This means not just changing the models but updating their actions as well.

This change also touched the Query, Url and RequestWrapper classes in order to process non-location based models.