Made a few changes to the Hook system. There are two new features-
Hook->loadModelPlugins takes a model class as an argument and generates the realm/category based off of that. It also loads plugins from the model type of any classes the passed model descends from, as long as that plugin allows it. Plugins with the model type "All" will be inherited by everything.

Hook::registerModelPlugin registers plugins designed for models. This is just a wrapper around the base registerPlugin function, more for connivence than functionality.