Big update! Running this revision directly will probably not work, as I am making a series of commits broken up more for ease of reading than anything else.
This update includes the long awaited separation of the Model type into a base Model type and and extension based on the location system, aptly named LocationModel. LocationModels are closer to the current previous Models in functionality, although that functionality has been improved. Models themselves are a bit more barebones, but they allow us to put some key system datatypes, such as Users, under the model system.

All of the base Model actions should be considered pre-Alpha at this point.

The location Read and Index actions for the LocationModels have advanced quite a bit. The Read class has a bunch of helper functions for converting models to html and create navigation menus, which the Index class builds off of to create tables of models that are children to the active models. 

The edit, add and delete classes still need to be refined.