Expanded the Table class.
Added the ability to repeat the headers every x number of rows, and to specify a minimum number of remaining rows for headers to be repeated (thus preventing headers to be repeated for only 1 or 2 rows).

Added an 'index' feature, which creates a column and numbers each row.

Added the ability to add CSS classes to the table.