Added documentation to the AbstractModel and LocationModel classes.
Updated the AbstractModel's checkAuth function. Originally this function was not defined, instead waiting for a custom definition from extending classes, but now it is set to check permissions at the root location using the model type and action. This will most likely be changed in a lot of custom classes but acts as a good starter point.

Updated the AbstractModel's delete function to return a boolean when saved.

Turned some of the AbstractModel's path functions into static functions.

Rewrote the LocationModel's delete function to better handle deleting child locations and errors.

Updated the LocationModels checkAuth function, which previously wasn't used and was directly inherited from the AbstractModel, to manage permissions for users attempting to perform actions on said model.

Updated the Add, Base and Edit actions to ask the models for permission, instead of doing all the work themselves. They still have the checkAuth function, which allows situations like the Add action needing to check the parent model for permission, while still allowing us to define the actual permission scheme in the models themselves.