Created a new folder for storing third party files, as well as an index to describe where they came from and various other data.
Updated the autoloader to crawl the third party directory.

Updated the installer to set the thirdparty path in the configuration file.

Updated the Config class to load the third party path in its fallback method.

Added the cssmin and JSMin classes to the third party collection.

Created the Minifier class as a wrapper around the cssmin and jsmin classes.