Updated the documentation of the base locationbased action.
Updated the Theme class to supply links to the new combined/minified css and javascript. Also updated the Theme class to support priorities when loading file paths and to return a primed minification object when needed.

Updated the Page class to use the new minified javascript instead of calling each script separately.