Made a few performance tweaks to the Http processor, Page class, theme and minify action.
Added a flush statement directly after the echo on the Http processor (this is probably next to useless, but I'm interested).

Moved the script includes, startup script and direct script to the bottom of the page by splitting the 'headerTemplate' into a separate 'footerTemplate' and appending that new string to the bottom of the template upon creation. This has resulted in a noticeable increase in page load speed, especially when the javascript has to be minified due to a cache hit (that can take upwards of 15 seconds but no longer stops the page from slowing down).

Modified the theme class to cache the javascript and css url so that it would not need to rely on the minifiers (rather expensive, since it requires opening a bunch of files) checksum function.

Modified the Minify action to invalidate the theme's url cache when the minified file changes. Updated documentation and increased the cache time of minified files considerably (since they invalidate based off of the checksum anyways).