Added table name to model listings generated from the ModelLocation Index action.
Changed the order that the LocationBased base action set its menu and loads its plugins so that the menu is properly set up for the plugin.

Created the PageHistory class. This action looks through the history of revisions and displays them in a list so that people can view them individually. It gets inherited by the blog module, which is nice.

Added a plugin to place a 'history' link on all of the Page classes and its submodels. Updated the installer to register the plugin at start.