Renovated the Installer a bit-
Split it out of the BentoBase module and into its own "installer" module. This is primarily to allow us to add more installation specific actions.

Added a new intercept for the javascript theme stuff- basically, since the base module isn't installed we need to intercept the theme call to it for loading javascript and dump it out ourselves. We also don't want to minify it here because caching isn't available and it will have to be redone each time.