Updated the Cache class to give it a "memory only" setting. When this setting is invoked (via the new 'setMemOnly' function) the cached data is kept in the memStore for the duration of the script but is sent or retrieved from the cache handler (sql, filesystem).
Added code to the Cache class to make sure it empties the $memStore when the Clear function is called.

Updated the PackageInfo class to return false when meta values aren't present. Also updated it to use the new 'memOnly' cache setting for when it loads the meta data associated with a package.

Corrected a bug in the Moduleinstaller that resulted in a bunch of notices being thrown there is no value set for the current package.