Updated the FormToHtml class to clean up the generation of html tag ids and to simplify the html generation.
Updated the Form class to remove all of the Html generation code that is now being used by the FormToHtml class. Added a couple of methods to the Form class that were needed by the FormToHtml class, mostly just data retrieval stuff.

Updated the HtmlObject class to have a list of tags which do not, by default, have closing tags (such as br and hr).