Mysql - indented the MysqlBase connection info.
ActionBase, ModelActionBase - set the permissions object to load off of the root location, not the active site.

Altered the Query class to no longer set the iohandler for the request wrapper, but rather store the iohandler in the array it generates. The default formats are now also set by the Query class.