Committing the new Cron handler.
Added the Cron handler itself, which is an ioProcessor that plugs into the RequestWrapper. 
The Cron handler takes advantage of the "nextRequest" function to run multiple actions from the cron queue.
The cron handler also uses a pid file which can be deleted to signal the request shut down.
When the cron request is told to shut down it will any running actions, The system is smart enough to detect that an action is running and check to make sure the process didn't die. If it died then CronStart script clean up the request as best it can so the action can be run again.

Added a start and stop action for the cron handler.
Added a new argument to the Argv handler to make calling cron easier (invoke the index.php file with the -r flag).
Updated the database to hold a cron table for storing actions and their time restrictions.