Updated the Form system.
Made it so default values can be set for password type fields without sending those defaults to the client. In otherwords, if the client leaves a password field blank but there is already a password for that object, then the password will stay the same.

Improved the error handling for when a value fails validation.

Corrected a bug in the Minimum Length and Maximum Length validators that was causing it not to validate properly (always failed).

Updated the ValidationLookup class, as it was pointing a rule at the wrong class.