Updated the Autocomplete plugin to the latest version and added the ajaxQueue and thickbox plugins.
Updated the Mortar form plugin to extend and enhance the Autocomplete plugin. This extension includes using Json as a data source and pulling individual settings for the autosuggest out of the metadata stored in the individual inputs.

The mortar form plugin update also includes a fix to the fck editor issue, where it would not validate properly unless the form was submitted multiple times.