Added a UserLookUp action which can return results in json.
Added two plugins to the Form class to create custom inputs, "user" and "location". These custom inputs utilize the autocomplete plugins to assist the user in filling in values. Only the user input is finished, as the location needs a backend lookup class still.

Updated the CustomInstall plugin to install the two new plugins during installation.