Added a new Hook to the form class inside the "checkSubmit" function and added documentation for both that hook and the one in the FormToHtml class.
Updated the Form class to allow autosuggest fields to return multiple results (still needs work, to convert the response back into an array from a string automatically).

Extended the FormInputUserToHtml plugin to convert an array of users into a string for the client side.

Added a plugin to convert the string returned by a User field into a user id or, if the 'multiple' option is selected an array of user ids.

Updated the Mortar module custom installer to register the new plugin.