Updated the Session Manager code.
When UserAgent checks don't match up the code now checks to see if the session was started with, and is still running, IE8. This is because IE8 will change its useragent for compatibility mode, but it still uses a specific tag to mark that it is IE8.

Updated the IP address checking to allow AOL users to switch from one proxy to another.