Cleaned out some code from the PackageInfo class which is no longer needed. Specifically the code loaded classes into the environment, a job which no is passed to the AutoLoader.
Had the PackageInfo class load its own module into the AutoLoader when called so it would have access to its files. This meant updating the AutoLoader so it wouldn't attempt to load the module multiple times (since it uses the PackageInfo class to get path data it would have caused a loop).

Updated the staticHack function (uug, can't wait for php 5.3 to be more widely adopted) to use the autoloader when the class isn't available.