Made another round of bug fixing using that same large data set.
At this point the stemmer works for every example given in the example dictionary except one, the word Fluently. However, both "fluently" and "fluent" use the same stem in this implementation- "fluent"- rather than using "fluent" and "fluenti" as they do in the original, so this is actually a small improvement.