Updated the installer to only purge the database if told to do so by the user installing it.
This update also makes it possible to clear out the database before returning with an error, so the installation process can continue normally. In other words, developers will be able to just clear out the configuration files, set a form field, and the installer will clear out the database for them.

(fixed issue 18)