This update extends the "theme inheritance" functionality to images using Twig.
Updated the Theme class to provide caching for image lookups and to pre-process the CSS and JS files with Twig.

Created a "ThemeImageWrapper" class for use with Twig to give designers access to the image information. By using the "images" variable they can call images up (images['logo.png']) and even test to see if they exist.

Updated the Page class to also use Twig and the ThemeImageWrapper.

Updated the cloisonne index.html file to use the new feature.

Removed whitespace from the minify action.