Adjusted the Cache class to expose some encoding/decoding functions.
Adjusted the FileSystem handler so that it included its php file rather than loading it as a string and parsing it. The decoding techniques are now directly embedded into the php class for better performance (particularly with scalars).

Adjusted the SQLite handler to better deal with encoding of objects, which requires storing the encoding used. Like the with the filesystem handler the goal of this is to raise performance when storing items that don't need to be serialized.