Organized the index.php file into a "BootStrapper" class, rather than just a long script. This is mostly for organizational purposes and to make it easier to follow the code.
Added a new constant, BASE_URL. This is the domain plus any 'real' path data before the virtual directories or dispatcher is shown, but not the protocol aspects (like http://).

Tweaked the "register_globals" check to be a bit more specific as to when it runs- previously it would run if "register_globals" was true, even though the value "off" converts to the boolean true.