Updated the Url subsystem to better handle a number of situations better when creating/reading the "virtual paths" in Urls.
This includes a lot of removing of defaults (such like the "http" iohandler and the default actions for modules/locations). 

Picked off some values that were being added to the query data rather than the path.

Cleaned up the UrlWriter by adding specific functions to handle each type of path to shrink down the size of the "makePath" function.