Updated the Caching class to fix a bug and optimize the way minification occurs.
The first bug was a case where infinite redirects could occur for intermittent periods due in part to a caching issue. To resolve this I placed a check in the minification code to make sure redirects would not happen if the redirect url was the same as the current url.

Used the new ProcessPool class to make it so only one minification (per file type, currently js or css) would occur at a time. In the event that one process is already handling minification the action will now output the raw javascript rather than starting an additional minification cycle. This is really fun to test- just clear your cache, start the js load in one window where it will run the minification code and then run it again in a second window. You can hit refresh in that second window a few dozen times with no issue while the first is minifiying.