Added a new function, "define", to the BootStapper class which sets constants if they have not been set already and updated the current profiles to use this function.
The reason for doing this is to allow system admins to override settings without having to modify the built in profiles. Currently admins have to use the profiles, edit them, or bypass them completely. This is no good from an updating perspective, because it means we either have to overwrite the admin's changes or we can't rely on new constants getting added to the system. This will also lead to a new set of constants (paths and other OS profiles) which will be overwritable by the sys admin.