Updated the FileSystem handler for Stash- yet again- to further improve how it handles large caches. I decided to go in favor of less folders and a smaller limit (still over 8million, a ridiculous amount for the filesystem engine) as a balance against the overhead of purging the cache with so many folders. I also ditched the number specific processing in order to ensure an even distribution.
A new feature in this update is the ability to tag key elements as non-dynamic categories by adding two underscores at the beginning of the name. For example, getCache("__models", $type, $id, "__info); would tell the system that models and info are non-dynamic keys, which backend storage systems can use for optimization. At this point the FileSystem class uses this to avoid chunking categories into multiple folders, since the constant being passed would always be the same folders and provide no distribution benefits.